Billing fixed Firstly a StartUpBillingUseCase has been added to allow the billing sdk to load
Secondly
we use a MutableShared state to observe the billing updates.
The reason for this was because when we made a pruchase, the event
NewPurchasesAvailable was fired but when another purchase was made again
the same NewPurchasesAvailable would not send with a mutable state,
since this state NewPurchasesAvailable would have been the same as the
previous one NewPurchasesAvailable.
Maybe I could have added the data to accompany NewPurchasesAvailable,
to make it unique.
I also added a refresh, to trigger a queryAsync products on the billing
lib, which in turn triggers listeners to update whoever is listening